requirements.txt
google-generativeai
python-dotenv
flask
flask-socketio
eventlet
pexpect
